backend-flask/requirements.txt
openai>=1.76.0
langchain>=0.3.24
langchain-core>=0.3.56
faiss-cpu>=1.10.0
flask>=3.0.2
flask-cors>=5.0.1
python-dotenv>=1.1.10
pypdf>=5.4.0
tiktoken>=0.9.0